Refactors selectors.js -> rename to general-selectors.js
