+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/chat-textfield-simple.js
@@ -22,7 +22,7 @@ import {
   selectProposableMessagesByConnectionUri,
   selectClaimableMessagesByConnectionUri,
   selectSelectedMessagesByConnectionUri,
-} from "../selectors/selectors.js";
+} from "../selectors/general-selectors.js";
 import { getAllMessageDetails } from "../won-utils.js";
 import autoresizingTextareaModule from "../directives/textarea-autogrow.js";
 import { actionCreators } from "../actions/actions.js";
