+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/feedback-grid.js
@@ -4,7 +4,7 @@ import "ng-redux";
 import { attach } from "../utils.js";
 import { actionCreators } from "../actions/actions.js";
 import { connect2Redux } from "../won-utils.js";
-import { selectIsConnected } from "../selectors/selectors.js";
+import { selectIsConnected } from "../selectors/general-selectors.js";
 
 import "style/_feedback-grid.scss";
