+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/messages/connection-message-actions.js
@@ -5,7 +5,7 @@ import Immutable from "immutable";
 import { connect2Redux } from "../../won-utils.js";
 import { attach, getIn } from "../../utils.js";
 import { actionCreators } from "../../actions/actions.js";
-import { selectNeedByConnectionUri } from "../../selectors/selectors.js";
+import { selectNeedByConnectionUri } from "../../selectors/general-selectors.js";
 import {
   isMessageProposable,
   isMessageClaimable,
