+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/messages/connection-message-status.js
@@ -5,7 +5,7 @@ import { relativeTime } from "../../won-label-utils.js";
 import { connect2Redux } from "../../won-utils.js";
 import { attach, getIn } from "../../utils.js";
 import { actionCreators } from "../../actions/actions.js";
-import { selectNeedByConnectionUri } from "../../selectors/selectors.js";
+import { selectNeedByConnectionUri } from "../../selectors/general-selectors.js";
 
 import "style/_connection-message-status.scss";
