+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/messages/connection-message.js
@@ -13,7 +13,7 @@ import combinedMessageContentModule from "./combined-message-content.js";
 import { connect2Redux } from "../../won-utils.js";
 import { attach, getIn } from "../../utils.js";
 import { actionCreators } from "../../actions/actions.js";
-import { selectNeedByConnectionUri } from "../../selectors/selectors.js";
+import { selectNeedByConnectionUri } from "../../selectors/general-selectors.js";
 import {
   isMessageProposable,
   isMessageClaimable,
