+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/messages/referenced-message-content.js
@@ -5,7 +5,7 @@ import won from "../../won-es6.js";
 import { connect2Redux } from "../../won-utils.js";
 import { attach, getIn } from "../../utils.js";
 import { actionCreators } from "../../actions/actions.js";
-import { selectNeedByConnectionUri } from "../../selectors/selectors.js";
+import { selectNeedByConnectionUri } from "../../selectors/general-selectors.js";
 import { labels } from "../../won-label-utils.js";
 import { fetchMessage } from "../../won-message-utils.js";
 import { classOnComponentRoot } from "../../cstm-ng-utils.js";
