+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/post-content-general.js
@@ -8,7 +8,7 @@ import { attach } from "../utils.js";
 import won from "../won-es6.js";
 import { labels, relativeTime } from "../won-label-utils.js";
 import { connect2Redux } from "../won-utils.js";
-import { selectLastUpdateTime } from "../selectors/selectors.js";
+import { selectLastUpdateTime } from "../selectors/general-selectors.js";
 import { actionCreators } from "../actions/actions.js";
 import ratingView from "./rating-view.js";
