+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/post-header.js
@@ -9,7 +9,7 @@ import { actionCreators } from "../actions/actions.js";
 import { labels, relativeTime } from "../won-label-utils.js";
 import { attach } from "../utils.js";
 import { connect2Redux } from "../won-utils.js";
-import { selectLastUpdateTime } from "../selectors/selectors.js";
+import { selectLastUpdateTime } from "../selectors/general-selectors.js";
 import won from "../won-es6.js";
 import { classOnComponentRoot } from "../cstm-ng-utils.js";
