+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/post-messages.js
@@ -24,7 +24,7 @@ import {
   selectCancellationPendingMessagesByConnectionUri,
   selectProposalMessagesByConnectionUri,
   selectUnreadMessagesByConnectionUri,
-} from "../selectors/selectors.js";
+} from "../selectors/general-selectors.js";
 import autoresizingTextareaModule from "../directives/textarea-autogrow.js";
 import { classOnComponentRoot } from "../cstm-ng-utils.js";
