+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/settings/settings-wrapper.js
@@ -2,7 +2,7 @@ import angular from "angular";
 import { Elm } from "../../../elm/Settings/Personas.elm";
 import { actionCreators } from "../../actions/actions";
 import "../identicon.js";
-import { getPersonas, currentSkin } from "../../selectors/selectors.js";
+import { getPersonas, currentSkin } from "../../selectors/general-selectors.js";
 
 function genComponentConf($ngRedux) {
   return {
