+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/topnav.js
@@ -12,7 +12,7 @@ import modalDialogModule from "./modal-dialog.js";
 import { attach, getIn } from "../utils.js";
 import { actionCreators } from "../actions/actions.js";
 import { connect2Redux } from "../won-utils.js";
-import { selectNeedByConnectionUri } from "../selectors/selectors.js";
+import { selectNeedByConnectionUri } from "../selectors/general-selectors.js";
 import "angular-marked";
 
 import * as srefUtils from "../sref-utils.js";
