+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/usecase-picker.js
@@ -9,7 +9,7 @@ import "ng-redux";
 import { attach, getIn } from "../utils.js";
 import { actionCreators } from "../actions/actions.js";
 import { connect2Redux } from "../won-utils.js";
-import { selectIsConnected } from "../selectors/selectors.js";
+import { selectIsConnected } from "../selectors/general-selectors.js";
 import { useCases, useCaseGroups } from "useCaseDefinitions";
 
 import "style/_usecase-picker.scss";
