+++ webofneeds/won-owner-webapp/src/main/webapp/app/configRouting.js
@@ -9,7 +9,7 @@ import { accountLogin } from "./actions/account-actions.js";
 
 import { privateId2Credentials } from "./won-utils.js";
 
-import { selectAllNeeds } from "./selectors/selectors.js";
+import { selectAllNeeds } from "./selectors/general-selectors.js";
 
 import {
   decodeUriComponentProperly,
